Separated masonry gallery files
