+++ classes/class-uagb-init-blocks.php
@@ -381,8 +381,8 @@ class UAGB_Init_Blocks {
 				'uagb_url'               => UAGB_URL,
 				'uagb_mime_type'         => UAGB_Helper::get_mime_type(),
 				'uagb_site_url'          => UAGB_URI,
-				'uagb_display_condition' => apply_filters( 'enable_block_condition', true ),
-				'enableMasonryGallery'   => apply_filters( 'enable_masonry_gallery', true ),
+				'uagb_display_condition' => apply_filters( 'uag_enable_block_condition', true ),
+				'enableMasonryGallery'   => apply_filters( 'uag_enable_masonry_gallery', true ),
 			)
 		);
