+++ src/blocks/extensions/block.js
@@ -1,8 +1,8 @@
+import './masonry-gallery';
 import UserConditionOptions from './condition-block';
 
 import { __ } from '@wordpress/i18n';
 const enableConditions = uagb_blocks_info.uagb_display_condition;
-const enableMasonryGallery = uagb_blocks_info.enableMasonryGallery;
 
 const AdvancedControlsBlock = wp.compose.createHigherOrderComponent((BlockEdit) => {
 
@@ -30,39 +30,6 @@ const AdvancedControlsBlock = wp.compose.createHigherOrderComponent((BlockEdit) 
 		);
 	};
 }, 'AdvancedControlsBlock');
-
-const MasonryGallery = wp.compose.createHigherOrderComponent((BlockEdit) => {
-
-	return (props) => {
-		
-		const { Fragment } = wp.element;
-		const { attributes, setAttributes, isSelected } = props;
-		const { InspectorAdvancedControls } = wp.blockEditor;
-		const {
-			ToggleControl
-		} = wp.components;
-		const blocks_name = props.name;
-		const block_type = [ 'core/gallery' ];
-		return (
-			<Fragment>
-				<BlockEdit {...props} />
-				{isSelected && block_type.includes(blocks_name) &&
-					<InspectorAdvancedControls>
-						<hr />
-						<p><strong>{ __( "Masonry Gallery", 'ultimate-addons-for-gutenberg' ) }</strong></p>
-						<p className="components-base-control__help">{ __( "Below setting will only take effect once you are on the live page, and not while you're editing.", 'ultimate-addons-for-gutenberg' ) }</p> 
-						<ToggleControl
-							label={ __('Enable Masonry Layout')}
-							checked={attributes.masonry}
-							onChange={(value) => setAttributes({ masonry: !attributes.masonry })}
-						/>
-						<hr />			
-					</InspectorAdvancedControls>
-				}
-			</Fragment>
-		);
-	};
-}, 'MasonryGallery');
 
 function ApplyExtraClass(extraProps, blockType, attributes) {
 
@@ -99,13 +66,7 @@ wp.hooks.addFilter(
 	'uagb/apply-extra-class',
 	ApplyExtraClass,
 );
-if ( enableMasonryGallery ) {
-	wp.hooks.addFilter(
-		'editor.BlockEdit',
-		'uagb/masonry-gallery',
-		MasonryGallery,
-	);
-}
+
 if(enableConditions){
 	wp.hooks.addFilter(
 		'editor.BlockEdit',
